fixed admin menu problems
