+++ app/src/main/AndroidManifest.xml
@@ -33,6 +33,7 @@
         <activity android:name="com.example.carbon_project.Controller.EntrantDashboardActivity" />
         <activity android:name="com.example.carbon_project.Controller.EntrantEventsListActivity" />
         <activity android:name="com.example.carbon_project.Controller.ProfileActivity" />
+        <activity android:name="com.example.carbon_project.Controller.AdminListActivity"/>
 
     </application>
