+++ app/src/main/java/com/example/carbon_project/Controller/AdminFacilityListFragment.java
@@ -1,4 +1,4 @@
-package com.example.carbon_project;
+package com.example.carbon_project.Controller;
 
 import android.os.Bundle;
 import androidx.annotation.NonNull;
@@ -10,23 +10,26 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import com.example.carbon_project.Model.Facility;
+import com.example.carbon_project.R;
+
 import java.util.ArrayList;
 import java.util.List;
 
-public class AdminEventListFragment extends Fragment {
+public class AdminFacilityListFragment extends Fragment {
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_admin_event_list, container, false);
+        View view = inflater.inflate(R.layout.fragment_admin_facility_list, container, false);
 
-        RecyclerView recyclerView = view.findViewById(R.id.recycler_view_tab2);
+        RecyclerView recyclerView = view.findViewById(R.id.recycler_view_tab3);
         recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
 
         List<Object> data = new ArrayList<>();
-        data.add(new Event("Event 1", "123 Apple Way", 23, false, "April 1", "April 4", "Test event", 30, "Test URL"));
-        data.add(new Event("Event 2", "123 Apple Way", 23, false, "April 1", "April 4", "Test event", 30, "Test URL"));
-        data.add(new Event("Event 3", "123 Apple Way", 23, false, "April 1", "April 4", "Test event", 30, "Test URL"));
+        data.add(new Facility("facilityID1", "Facility 1", "Anderossa", 500, "False_Org"));
+        data.add(new Facility("facilityID2", "Facility 2", "Anderossa", 500, "Falser_Org"));
+        data.add(new Facility("facilityID3", "Facility 3", "Anderossa", 500, "Falsest_Org"));
 
         AdminRecyclerViewAdapter adapter = new AdminRecyclerViewAdapter(data);
         recyclerView.setAdapter(adapter);
